# assets/mes/services.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "services.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/mes/api/v2/resource/collect": {
      "post": {
        "summary": "收料",
        "description": "收料功能將會進行收料以及產出紀錄. 您必須指定 workOrder, collectBatch, quantity 告訴系統欲收料的工單號碼, 收料量\n相當於幾個首數(預設為1), 收料量.\n部份工程不提供工單要可以進行收料, 則需指定 product 資訊(id, type, unit)\nps: 不適用於包含投料功能的收料\n\n您可透過 resource 指定這批產出的製品/成品紀錄至指定的標示卡條碼, 若無指定, 系統將自動產生一組標示卡條碼, 可從\nCollectReply.resource 得知其標示卡號碼. 當您未指定標示卡條碼時, 您必須透過 carrier 指定其標示卡條碼應放置於哪一\n載具.\n部份工程沒有使用載具或是沒有規劃相關流程, 但有指定收料條碼(resource.ID), 則 carrier 可以不傳.\n\n系統依據 actionMode 分成兩種執行模式: 一般執行(ACTION_AUTO)和強制執行(ACTION_FORCE), 預設為一般執行. 當處理請求\n時遇到系統認定為不正常操作(例如: 該首數狀態不可收料, 收料量大於投料量等)時, 會回饋錯誤到 CollectReply.error.\n您亦可透過強制執行(actionMode=ACTION_FORCE), 系統仍回饋錯誤並執行收料及產出紀錄, 依情況恐怕發生資料不齊全或無法\n預期狀況, 例如標示卡不存在的話, 將導致後段工程仍舊無法使用該條碼, 即使收料請求已被系統接受並執行.\n\n若您想知道該批製品/成品更進一步資訊, 可透過 labelFields 指定需要知道的資訊, 目前支援:\n - product_type: 製品/成品類別, 例如: 天然橡膠, 合成橡膠, 鋼絲圈...等.\n - process_name: 製品生產流程代號\n - collect_quantity: 收取數\n - station: 機台號碼\n - manufacture_date: 製造日期時間\n - expiry: 製品/成品過期日期時間\n - operator: 收料人員\n - batch: 工單號碼+collectBatch\n - lot: 標示卡號碼\n - lot_number: 生產批號, 格式: 1碼工作群組+2碼機台編號+\"-\"+2碼月份+2碼日期, 例如: 工作群組=5, 機台編號=3,\n               日期=2021-02-03, 其生產批號為: 503-0203.\n以上資訊可從 CollectReply.labelFields 獲得.\n\n若您僅需要查詢收料紀錄, 可指定 onlyQueryLabels=true, 並給予以下資訊:\n - 工單編號(workOrder)\n - 收料序號(sequence)\n - 查詢欄位(labelFields)\n系統亦依據 metadata 的操作機台和當前登入人員, 篩選對應的收料紀錄(依據工單編號+生產批號+收料序號).\n若存在收料記錄, 依據 labelFields 回傳資料; 若不存在則回傳錯誤.\n\n包含投料功能:\n - Feed.feeds 參考 resourceFeed 指定投入材料 or 從 site 投入指定材料\n - Feed.AccordingRecipe 決定是否依照配合表重量扣重, 不指定的話 feeds 投入材料必填重量\n\n下列為回傳錯誤清單:\n - Error_ERROR_BAD_REQUEST\n - ERROR_RESOURCE_NOT_FOUND (材料條碼不存在)\n - ERROR_RESOURCE_EXISTED (收料標示卡已存在)\n - ERROR_WORKORDER_NOT_FOUND (找不到指定的工單)\n - ERROR_WORKORDER_BAD_BATCH (該工單作業批次狀態無法進行掛載)\n - ERROR_CARRIER_NOT_FOUND (找不到指定收取容器條碼)\n - ERROR_CARRIER_IN_USE (收料容器已被使用)\n - ERROR_BATCH_NOT_READY (該作業批次還不能收料)\n - ERROR_RW_QUANTITY_ABOVE_MAX (收取數量大於投料量, 超出允許收料範圍)\n - ERROR_WORKORDER_BAD_STATUS (工單狀態錯誤不允許進行操作)\n - ERROR_RECORD_EXISTED (收料紀錄已經存在)\n - ERROR_RECORD_NOT_FOUND (找不到收料紀錄)",
        "operationId": "Collect",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiCollectReply"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiCollectRequest"
            }
          }
        ],
        "tags": [
          "API"
        ]
      }
    },
...
    "/mes/api/v2/workorder": {
      "post": {
        "summary": "工單操作",
        "description": "工單操作包含以下功能:\n - 變更工單狀態(工單狀態請參考 ./workorder/workorder.proto enum Status),\n   包含以下變更方式:\n    - 跳過工單: 將 skipWorkOrders 內所有工單狀態變更為 SKIPPED. 將工單狀態從\n               PENDING -\u003e SKIPPED, 其他狀態不可執行此操作.\n    - 投料結束工單: closeWorkOrder=指定結束工單, 工單狀態必須為 ACTIVE,\n                   同時所有的首數(batch)狀態為 BATCH_PREPARING / BATCH_CLOSING / BATCH_CLOSED /\n                   BATCH_CANCELLED (首數狀態請參考 ./workorder/workorder.proto enum BatchStatus)\n                   任一種狀態, 系統將變更工單狀態為 CLOSING, 否則不可執行操作並將錯誤回饋到\n                   WorkOrderReply.error.\n                   系統一併會把全部首數狀態 = BATCH_PREPARING 更新成 BATCH_CANCELLED 代表該首數已被取消.\n      您亦可透過指定 closeMode=ACTION_FORCE 進行強制執行, 程式同樣回饋錯誤警示並變更工單狀態.\n      若工單狀態非上述允許的狀態, 則不可執行此操作, 即使指定 closeMode=ACTION_FORCE 也無法變更狀態.\n    - 收料結束工單: finishWorkOrder=指定結束工單, 不檢查首數狀態.\n    - 執行工單: 依據 startWorkOrder 執行指定工單. 將工單狀態從 PENDING -\u003e ACTIVE, 其他狀態不可執行此操作.\n - 查詢工單資料: 依據 loadWorkOrders 查詢工單資料, 詳細操作請參考 ./workorder/workorder.proto\n               message WorkOrderRequest.LoadWorkOrders. 其查詢結果為 workOrders.\n    - 查詢工單(todosWithDetails): 表示取得工單需含該作業流程所需用到配方以及機台參數.\n - 變更首數狀態: 透過指定 batchAction 變更指定工單的首數狀態.\n\n若一個請求包含上述所有功能, 系統執行順序為 [變更首數狀態] -\u003e [變更工單狀態] -\u003e [查詢工單資料].\n\n取得TODO工單回傳資訊:\n - id: 工單號\n - process: 生產作業階段\n - batches.count: 目標生產總批次\n - batches size分三種類型表示(只會出現一種), 分別用來表示 batches.count 對應關係:\n    - unitQuantity: 為 true 的話, 表示為單位數量, 如: 成型每次生產一條生胎\n    - fixedQuantity: 每次批次產出數量為固定數值, 依研發單位制定\n    - perBatchQuantities: 每次批次產出數量不同, 派工決定該批次產出數量\n - status: 工單狀態, 請參考 ./workorder/workorder.proto enum Status\n - fedBatch: 已經投入材料的批次\n - collectedBatch: 已收取成品的批次\n - recipe: details格式為[key]value的型式, 如下說明:\n    - groupedResources : 材料\n       - STEP-${n}-MATERIALS : 表示各階段所需使用的材料, ${n}為大於0整數\n          - requirements : 材料清單\n       - TOOLS : 表示作業所需使用的製工具清單\n          - requirements : 工具清單\n    - groupedParams : 機台參數\n       - STEP-${n}-CONTROL : 表示各階段機台設定參數, ${n}為大於0整數\n          - params : 參數清單\n       - STEP-${n}-MEASUREMENT : 表示各階段機台量測參數, ${n}為大於0整數\n          - params : 參數清單\n\n下列為回傳錯誤清單:\n - Error_ERROR_BAD_REQUEST\n - Error_ERROR_BATCH_NOT_FOUND (找不到工單作業批次)\n - Error_ERROR_WORKORDER_BAD_BATCH (錯誤的作業批次)\n - Error_ERROR_WORKORDER_NOT_FOUND (找不到工單)\n - Error_ERROR_WORKORDER_BAD_STATUS (工單狀態錯誤不允許進行操作)",
        "operationId": "WorkOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiWorkOrderReply"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiWorkOrderRequest"
            }
          }
        ],
        "tags": [
          "API"
        ]
      }
    }
  },
  "definitions": {
...
    "CollectReplylabelField": {
      "type": "object",
      "properties": {
        "time": {
          "$ref": "#/definitions/commonsTime"
        },
        "decimal": {
          "$ref": "#/definitions/v2commonsDecimal"
        },
        "string": {
          "type": "string"
        }
      }
    },
    "CollectRequestFeed": {
      "type": "object",
      "properties": {
        "feeds": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiResourceFeed"
          }
        },
        "accordingRecipe": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "CollectRequestProduct": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "unit": {
          "type": "string"
        }
      }
    },
    "ErrorCode": {
      "type": "string",
      "enum": [
        "ERROR_NONE",
        "ERROR_INTERNAL",
        "ERROR_BAD_REQUEST",
        "ERROR_USER_NOT_FOUND_OR_BAD_PASSWORD",
        "ERROR_USER_NO_PERMISSION",
        "ERROR_USER_MISSING_TOKEN",
        "ERROR_USER_UNKNOWN_TOKEN",
        "ERROR_STATION_NOT_FOUND",
        "ERROR_STATION_ON_HOLD",
        "ERROR_RESOURCE_NOT_FOUND",
        "ERROR_RESOURCE_EXPIRED",
        "ERROR_RESOURCE_ON_HOLD",
        "ERROR_RESOURCE_EXISTED",
        "ERROR_RESOURCE_CONTROL_ABOVE_EXTENDED_COUNT",
        "ERROR_WORKORDER_NOT_FOUND",
        "ERROR_WORKORDER_BAD_BATCH",
        "ERROR_WORKORDER_BAD_STATUS",
        "ERROR_WORKORDER_UNEXPECTED_RESULTS",
        "ERROR_CARRIER_NOT_FOUND",
        "ERROR_CARRIER_IN_USE",
        "ERROR_CARRIER_FULL",
        "ERROR_BATCH_NOT_FOUND",
        "ERROR_BATCH_NOT_READY",
        "ERROR_RECORD_NOT_FOUND",
        "ERROR_RECORD_EXISTED",
        "ERROR_RECIPE_NOT_FOUND",
        "ERROR_US_MISMATCH",
        "ERROR_US_SIGN_ON_FORBIDDEN",
        "ERROR_SR_RESOURCE_IN_USE",
        "ERROR_SR_SITE_NOT_FOUND",
        "ERROR_SW_MISMATCH",
        "ERROR_RW_RESOURCE_NOT_READY",
        "ERROR_RW_QUANTITY_BELOW_MIN",
        "ERROR_RW_QUANTITY_ABOVE_MAX",
        "ERROR_RW_QUANTITY_LOW_DOSAGE",
        "ERROR_RW_QUANTITY_SHORTAGE",
        "ERROR_RW_BAD_GRADE",
        "ERROR_RW_RESOURCE_UNEXPECTED",
        "ERROR_RW_RESOURCE_MISSING"
      ],
      "default": "ERROR_NONE",
      "title": "- ERROR_INTERNAL: code value \u003c 10000 reserved for plant side application, you should use gRPC error instead\nDO NOT USE IT AS MES API ERROR.\n - ERROR_STATION_ON_HOLD: e.g., maintenance.\n - ERROR_RESOURCE_EXISTED: resource id has been used and thus not valid for new resource from Collect\n - ERROR_WORKORDER_BAD_BATCH: the batch is not allowed to be operated. e.g. CLOSED batch can not be closed.\n - ERROR_WORKORDER_BAD_STATUS: work order unexpected status, e.g: load feed work order only allowed ACTIVE status, load collect work order only allowed ACTIVE or CLOSING status.\n - ERROR_WORKORDER_UNEXPECTED_RESULTS: work order unexpected results, e.g: requested todo work order for feeding but return zero-length todo reply.\n - ERROR_CARRIER_IN_USE: carrier in use and cannot be shared\n - ERROR_CARRIER_FULL: carrier can be shared but it's full\n - ERROR_BATCH_NOT_READY: not enough fed batch to collect\n - ERROR_RECORD_NOT_FOUND: not found query record.\n - ERROR_RECORD_EXISTED: record has been collected.\n - ERROR_US_MISMATCH: the user is not the station/site operator.\n - ERROR_US_SIGN_ON_FORBIDDEN: e.g., not qualified/registered to operate the station.\n - ERROR_SR_RESOURCE_IN_USE: the resource is bound to another station/site provided in details.\n - ERROR_SW_MISMATCH: the work order is not being executed the station\n - ERROR_RW_RESOURCE_NOT_READY: details: \"required_preprocess_name\", e.g., \"cutting\"\n - ERROR_RW_QUANTITY_BELOW_MIN: the used quantity is less than the minimum quantity in the recipe.\n - ERROR_RW_QUANTITY_ABOVE_MAX: the used quantity is greater than the maximum quantity in the recipe.\n - ERROR_RW_QUANTITY_LOW_DOSAGE: the used quantity is less than the dosage.\n - ERROR_RW_QUANTITY_SHORTAGE: the quantity of the stock is less than the specified quantity(e.g. used quantity, the minimum quantity in the recipe).\n - ERROR_RW_BAD_GRADE: details: \"required_resource_grade\", e.g., \"AA\"\n - ERROR_RW_RESOURCE_UNEXPECTED: the resource is not in the work order recipe\n - ERROR_RW_RESOURCE_MISSING: the resource is required but missing"
    },
...
    "WorkOrderReplyWorkOrders": {
      "type": "object",
      "properties": {
        "past": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v2workorderInfo"
          }
        },
        "todo": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v2workorderInfo"
          }
        }
      }
    },
    "WorkOrderRequestBatchAction": {
      "type": "object",
      "properties": {
        "workOrder": {
          "type": "string"
        },
        "batch": {
          "type": "integer",
          "format": "int64"
        },
        "status": {
          "$ref": "#/definitions/workorderBatchStatus"
        },
        "note": {
          "type": "string"
        }
      },
      "title": "Manual change to batch status"
    },
    "WorkOrderRequestLoadWorkOrders": {
      "type": "object",
      "properties": {
        "past": {
          "type": "integer",
          "format": "int64"
        },
        "todo": {
          "type": "integer",
          "format": "int64"
        },
        "pastIncludesSkipped": {
          "type": "boolean",
          "format": "boolean"
        },
        "pastIncludesClosed": {
          "type": "boolean",
          "format": "boolean"
        },
        "todosWithDetails": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "apiCollectReply": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/checkError"
        },
        "resource": {
          "$ref": "#/definitions/resourceID"
        },
        "labelFields": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/CollectReplylabelField"
          },
          "title": "more collected resource info for display and printing.\nstring表示標示卡欄位說明,labelField 為value"
        },
        "enforceable": {
          "type": "boolean",
          "format": "boolean"
        },
        "enforceDone": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "apiCollectRequest": {
      "type": "object",
      "properties": {
        "actionMode": {
          "$ref": "#/definitions/checkActionMode"
        },
        "workOrder": {
          "type": "string"
        },
        "collectBatch": {
          "type": "integer",
          "format": "int32"
        },
        "collector": {
          "type": "string"
        },
        "quantity": {
          "$ref": "#/definitions/v2commonsDecimal"
        },
        "resource": {
          "$ref": "#/definitions/resourceID"
        },
        "carrier": {
          "type": "string"
        },
        "labelFields": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "sequence": {
          "type": "integer",
          "format": "int32"
        },
        "onlyQueryLabels": {
          "type": "boolean",
          "format": "boolean"
        },
        "feed": {
          "$ref": "#/definitions/CollectRequestFeed"
        },
        "product": {
          "$ref": "#/definitions/CollectRequestProduct"
        }
      }
    },
...
    "apiResourceFeed": {
      "type": "object",
      "properties": {
        "site": {
          "$ref": "#/definitions/ResourceFeedSiteReference"
        },
        "resource": {
          "$ref": "#/definitions/apiResourceFeedResource"
        },
        "quantity": {
          "$ref": "#/definitions/v2commonsDecimal"
        },
        "cumulative": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
...
    "apiWorkOrderReply": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/checkError"
        },
        "workOrders": {
          "$ref": "#/definitions/WorkOrderReplyWorkOrders"
        }
      }
    },
    "apiWorkOrderRequest": {
      "type": "object",
      "properties": {
        "skipWorkOrders": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "closeWorkOrder": {
          "type": "string"
        },
        "startWorkOrder": {
          "type": "string"
        },
        "loadWorkOrders": {
          "$ref": "#/definitions/WorkOrderRequestLoadWorkOrders"
        },
        "note": {
          "type": "string"
        },
        "batchAction": {
          "$ref": "#/definitions/WorkOrderRequestBatchAction"
        },
        "closeMode": {
          "$ref": "#/definitions/checkActionMode"
        },
        "finishWorkOrder": {
          "type": "string"
        }
      }
    },
    "checkActionMode": {
      "type": "string",
      "enum": [
        "ACTION_AUTO",
        "ACTION_FORCE",
        "ACTION_PRECHECK"
      ],
      "default": "ACTION_AUTO"
    },
    "checkError": {
      "type": "object",
      "properties": {
        "code": {
          "$ref": "#/definitions/ErrorCode"
        },
        "details": {
          "type": "string"
        }
      }
    },
...
    "commonsTime": {
      "type": "object",
      "properties": {
        "nano": {
          "type": "string",
          "format": "int64",
          "title": "now.UnixNano()"
        }
      }
    },
...
    "resourceID": {
      "type": "object",
      "properties": {
        "type": {
          "$ref": "#/definitions/v2resourceType"
        },
        "id": {
          "type": "string"
        },
        "productType": {
          "type": "string"
        }
      }
    },
...
    "v2commonsDecimal": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "format": "int64"
        },
        "exp": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "v2resourceType": {
      "type": "string",
      "enum": [
        "MATERIAL",
        "TOOL"
      ],
      "default": "MATERIAL"
    },
    "v2workorderInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "process": {
          "type": "string"
        },
        "batches": {
          "$ref": "#/definitions/workorderBatches"
        },
        "recipe": {
          "$ref": "#/definitions/recipeRecipe"
        },
        "status": {
          "$ref": "#/definitions/workorderStatus"
        },
        "fedBatch": {
          "type": "integer",
          "format": "int32"
        },
        "collectedBatch": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
  }
}